+++ contracts/ClaimIssuer.sol
@@ -48,7 +48,7 @@ contract ClaimIssuer is IClaimIssuer, Identity {
         uint256 claimTopic,
         bytes memory sig,
         bytes memory data)
-    external override view returns (bool claimValid)
+    public override(Identity, IClaimIssuer) view returns (bool claimValid)
     {
         bytes32 dataHash = keccak256(abi.encode(_identity, claimTopic, data));
         // Use abi.encodePacked to concatenate the message prefix and the message to sign.
@@ -78,39 +78,5 @@ contract ClaimIssuer is IClaimIssuer, Identity {
         }
 
         return false;
-    }
-
-    /**
-     *  @dev See {IClaimIssuer-getRecoveredAddress}.
-     */
-    function getRecoveredAddress(bytes memory sig, bytes32 dataHash)
-        public override
-        pure
-        returns (address addr)
-    {
-        bytes32 ra;
-        bytes32 sa;
-        uint8 va;
-
-        // Check the signature length
-        if (sig.length != 65) {
-            return address(0);
-        }
-
-        // Divide the signature in r, s and v variables
-        // solhint-disable-next-line no-inline-assembly
-        assembly {
-            ra := mload(add(sig, 32))
-            sa := mload(add(sig, 64))
-            va := byte(0, mload(add(sig, 96)))
-        }
-
-        if (va < 27) {
-            va += 27;
-        }
-
-        address recoveredAddress = ecrecover(dataHash, va, ra, sa);
-
-        return (recoveredAddress);
     }
 }