+++ contracts/Identity.sol
@@ -493,6 +493,79 @@ contract Identity is Storage, IIdentity, Version {
     }
 
     /**
+     * @dev Checks if a claim is valid. Claims issued by the identity are self-attested claims. They do not have a
+     * built-in revocation mechanism and are considered valid as long as their signature is valid and they are still
+     * stored by the identity contract.
+     * @param _identity the identity contract related to the claim
+     * @param claimTopic the claim topic of the claim
+     * @param sig the signature of the claim
+     * @param data the data field of the claim
+     * @return claimValid true if the claim is valid, false otherwise
+     */
+    function isClaimValid(
+        IIdentity _identity,
+        uint256 claimTopic,
+        bytes memory sig,
+        bytes memory data)
+    public override virtual view returns (bool claimValid)
+    {
+        bytes32 dataHash = keccak256(abi.encode(_identity, claimTopic, data));
+        // Use abi.encodePacked to concatenate the message prefix and the message to sign.
+        bytes32 prefixedHash = keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", dataHash));
+
+        // Recover address of data signer
+        address recovered = getRecoveredAddress(sig, prefixedHash);
+
+        // Take hash of recovered address
+        bytes32 hashedAddr = keccak256(abi.encode(recovered));
+
+        // Does the trusted identifier have they key which signed the user's claim?
+        //  && (isClaimRevoked(_claimId) == false)
+        if (keyHasPurpose(hashedAddr, 3)) {
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * @dev returns the address that signed the given data
+     * @param sig the signature of the data
+     * @param dataHash the data that was signed
+     * returns the address that signed dataHash and created the signature sig
+     */
+    function getRecoveredAddress(bytes memory sig, bytes32 dataHash)
+    public
+    pure
+    returns (address addr)
+    {
+        bytes32 ra;
+        bytes32 sa;
+        uint8 va;
+
+        // Check the signature length
+        if (sig.length != 65) {
+            return address(0);
+        }
+
+        // Divide the signature in r, s and v variables
+        // solhint-disable-next-line no-inline-assembly
+        assembly {
+            ra := mload(add(sig, 32))
+            sa := mload(add(sig, 64))
+            va := byte(0, mload(add(sig, 96)))
+        }
+
+        if (va < 27) {
+            va += 27;
+        }
+
+        address recoveredAddress = ecrecover(dataHash, va, ra, sa);
+
+        return (recoveredAddress);
+    }
+
+    /**
      * @notice Initializer internal function for the Identity contract.
      *
      * @param initialManagementKey The ethereum address to be set as the management key of the ONCHAINID.